fix: bunch of bug fixes
