+++ wasm_module/src/data/character.rs
@@ -1,4 +1,6 @@
 #![allow(non_snake_case)]
+use crate::utils::serde_arrays;
+
 use super::{
     affix::Affix,
     attribute::{Attribute, ATTRIBUTE_COUNT},
@@ -42,19 +44,6 @@ pub struct Character {
     pub rankby: Attribute,
     pub gear: [Affix; 14],
     pub combination_id: u32,
-}
-mod serde_arrays {
-    use serde::Serialize;
-
-    use super::ATTRIBUTE_COUNT;
-
-    pub fn serialize<S, T>(array: &[T; ATTRIBUTE_COUNT], serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-        T: Serialize,
-    {
-        array[..].serialize(serializer)
-    }
 }
 
 impl Character {
@@ -108,104 +97,4 @@ impl Character {
 
         invalid
     }
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[repr(align(64))] // cache line alignment
-pub struct ResultCharacter {
-    // attributes indexed by Attribute enum
-    // array instead of struct since this is muuuuuuuuch faster than matching with enum
-    #[serde(with = "serde_arrays")]
-    pub base_attributes: Attributes,
-    #[serde(with = "serde_arrays")]
-    pub attributes: Attributes,
-
-    pub gear: [Affix; 14],
-    pub gear_stats: [f32; 10],
-    pub combination_id: u32,
-
-    pub results: Results,
-}
-impl ResultCharacter {
-    pub fn to_character(&self) -> Character {
-        Character {
-            base_attributes: self.base_attributes,
-            attributes: self.attributes,
-            rankby: Attribute::Power,
-            gear: self.gear,
-            combination_id: self.combination_id,
-        }
-    }
-    pub fn from(character: &Character) -> Self {
-        ResultCharacter {
-            base_attributes: character.base_attributes,
-            attributes: character.attributes,
-            gear: character.gear,
-            gear_stats: [0.0; 10],
-            combination_id: character.combination_id,
-            results: Results {
-                value: character.score(),
-                coefficientHelper: CoefficientHelper {
-                    Bleeding: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Burning: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Confusion: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Poison: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Power: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Power2: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                    Torment: CoefficientDetails {
-                        slope: 0.0,
-                        intercept: 0.0,
-                    },
-                },
-                damageBreakdown: [0.0; 7],
-                effectiveDamageDistribution: [0.0; 7],
-                effectiveNegativeValues: [0.0; 5],
-                effectivePositiveValues: [0.0; 5],
-            },
-        }
-    }
-}
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct Results {
-    pub coefficientHelper: CoefficientHelper,
-    pub damageBreakdown: [f32; 7],
-    pub effectiveDamageDistribution: [f32; 7],
-    pub effectiveNegativeValues: [f32; 5],
-    pub effectivePositiveValues: [f32; 5],
-    pub value: f32,
-}
-
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct CoefficientHelper {
-    pub Bleeding: CoefficientDetails,
-    pub Burning: CoefficientDetails,
-    pub Confusion: CoefficientDetails,
-    pub Poison: CoefficientDetails,
-    pub Power: CoefficientDetails,
-    pub Power2: CoefficientDetails,
-    pub Torment: CoefficientDetails,
-}
-
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct CoefficientDetails {
-    pub slope: f32,
-    pub intercept: f32,
 }
