+++ wasm_module/src/result.rs
@@ -1,15 +1,6 @@
+use crate::data::{character::Character, resultcharacter::ResultCharacter, settings::Settings};
 use serde::Serialize;
 use std::vec;
-
-use crate::{
-    data::{
-        attribute::Attribute,
-        character::{Character, ResultCharacter},
-        settings::Settings,
-    },
-    optimizer_core::update_attributes,
-};
-
 #[derive(Debug, Serialize)]
 pub struct Result {
     // this holds the finalized characters that will be returned to js
@@ -91,7 +82,8 @@ impl Result {
                     .for_each(|a| character.gear_stats[a.0 as usize] += a.1);
             }
 
-            calc_results(character, settings);
+            // fills the character.results object with additional details about the character
+            character.calc_results(settings);
         }
     }
 
@@ -116,67 +108,5 @@ impl Result {
             .for_each(|c| *c = (*c as f32 / sum as f32 * 100.0) as u32);
 
         combination_count
-    }
-}
-
-fn calc_results(character: &mut ResultCharacter, settings: &Settings) {
-    let attributes = character.attributes;
-    // results.value is assigned when calling ResultCharacter::from
-    // we skip indicators in rust to avoid redundancy
-
-    // baseline for comparing adding/subtracting +5 infusions
-    let mut baseline = character.clone().to_character();
-    update_attributes(&mut baseline, settings, true);
-
-    // effective gain/loss from +5 attributes
-    for (index, attribute) in [
-        Attribute::Power,
-        Attribute::Precision,
-        Attribute::Ferocity,
-        Attribute::ConditionDamage,
-        Attribute::Expertise,
-    ]
-    .iter()
-    .enumerate()
-    {
-        let mut copy = character.clone().to_character();
-        // add +5 infusions
-        copy.base_attributes[*attribute as usize] = copy.base_attributes[*attribute as usize] + 5.0;
-
-        update_attributes(&mut copy, settings, true);
-        character.results.effectivePositiveValues[index] = copy.attributes
-            [Attribute::Damage as usize]
-            - baseline.attributes[Attribute::Damage as usize];
-
-        // subtract -5 infusions
-        copy = character.clone().to_character();
-        copy.base_attributes[*attribute as usize] =
-            f32::max(copy.base_attributes[*attribute as usize] - 5.0, 0.0);
-
-        update_attributes(&mut copy, settings, true);
-        character.results.effectiveNegativeValues[index] = copy.attributes
-            [Attribute::Damage as usize]
-            - baseline.attributes[Attribute::Damage as usize];
-    }
-
-    for (index, attribute) in [
-        Attribute::PowerDPS,
-        Attribute::Power2DPS,
-        Attribute::BleedingDPS,
-        Attribute::BurningDPS,
-        Attribute::ConfusionDPS,
-        Attribute::PoisonDPS,
-        Attribute::TormentDPS,
-    ]
-    .iter()
-    .enumerate()
-    {
-        // effective damage distribution
-        character.results.effectiveDamageDistribution[index] =
-            character.attributes[*attribute as usize] / attributes[Attribute::Damage as usize]
-                * 100.0;
-
-        // damage breakdown
-        character.results.damageBreakdown[index] = character.attributes[*attribute as usize];
     }
 }
